+++ Components/ActionBars/Elements/StanceBar.lua
@@ -479,11 +479,11 @@ StanceBar.OnInitialize = function(self)
 end
 
 StanceBar.OnEnable = function(self)
-	if ns.callbacks and ns.callbacks.RegisterCallback then
+	if ns.RegisterCallback then
 		ns.RegisterCallback(self, "ActionBars_Artwork_Updated", "UpdatePosition")
 		ns.RegisterCallback(self, "ActionBars_SecondaryBar_Updated", "UpdatePosition")
 		ns.RegisterCallback(self, "ActionBars_ThirdBar_Updated", "UpdatePosition")
-		ns.callbacks:RegisterCallback(self, "StanceBar_Position_Updated", "UpdatePosition")
+		ns.RegisterCallback(self, "StanceBar_Position_Updated", "UpdatePosition")
 	end
 
 	self:RegisterEvent("PLAYER_ENTERING_WORLD", "OnEvent")